The Team/backend/api/serializers.py
from rest_framework import serializers
from .models import UserSubmission

class UserSubmissionSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserSubmission
        fields = ['id', 'full_name', 'university', 'photo', 'created_at']
