The Team/requirements.txt
django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
pillow==11.1.0
python-dotenv==1.0.1
dj-database-url==2.3.0
